# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAR Fleet Service",
    "summary": "Ground fleet service for civilian search-and-rescue drone operations.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "description": "List incidents, optionally by status.",
        "operationId": "list_incidents_api_v1_incidents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum items to return."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous page's next_cursor."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPage"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.view"
      },
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident",
        "description": "Open an incident.",
        "operationId": "create_incident_api_v1_incidents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "incidents.manage"
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "description": "One incident.",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.view"
      },
      "patch": {
        "tags": [
          "incidents"
        ],
        "summary": "Update Incident",
        "description": "Change an incident's details, operating area or status.",
        "operationId": "update_incident_api_v1_incidents__incident_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "incidents.manage"
      },
      "delete": {
        "tags": [
          "incidents"
        ],
        "summary": "Delete Incident",
        "description": "Delete an empty incident (for one opened by mistake); otherwise close it instead.",
        "operationId": "delete_incident_api_v1_incidents__incident_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "incidents.manage"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Where the value is: body, query, path, header."
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "description": "One invalid input value."
      },
      "GeoPoint": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "GeoPoint",
        "description": "A WGS84 position with explicit keys (never a bare pair)."
      },
...
      "IncidentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^\\S(?:.*\\S)?$",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "base": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "base_altitude_amsl_m": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 9000.0,
                "minimum": -500.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Altitude Amsl M"
          },
          "operating_radius_m": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Operating Radius M",
            "description": "Every geometry of the incident must lie within this distance of the base.",
            "default": 25000.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "base"
        ],
        "title": "IncidentCreate",
        "description": "A new incident; it starts ``active``."
      },
      "IncidentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "base": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "base_altitude_amsl_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Altitude Amsl M"
          },
          "operating_radius_m": {
            "type": "number",
            "title": "Operating Radius M"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "base",
          "base_altitude_amsl_m",
          "operating_radius_m",
          "created_by",
          "created_at",
          "updated_at",
          "closed_at"
        ],
        "title": "IncidentOut",
        "description": "A SAR incident."
      },
      "IncidentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Null when this is the last page."
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "IncidentPage",
        "description": "A page of incidents."
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "closed"
        ],
        "title": "IncidentStatus",
        "description": "Lifecycle of an incident; ``closed`` is terminal and read-only."
      },
      "IncidentUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^\\S(?:.*\\S)?$",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "base": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "base_altitude_amsl_m": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 9000.0,
                "minimum": -500.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Altitude Amsl M"
          },
          "operating_radius_m": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Operating Radius M",
            "description": "Every geometry of the incident must lie within this distance of the base."
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IncidentUpdate",
        "description": "Fields to change. ``closed`` is terminal; moving the base or shrinking the radius\nis refused while any geometry of the incident would end up outside."
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank",
            "examples": [
              "urn:sar-gcs:problem:conflict"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Present on validation errors: each invalid value."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details. Extension members depend on ``type``."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Session token returned by POST /api/v1/auth/login.",
        "scheme": "bearer"
      }
    }
  },
...
}